scripts/azure_func1/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
fastapi
uvicorn
pydantic
onnxruntime
Pillow
numpy
requests